% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_alignment_function.R
\name{show_alignment}
\alias{show_alignment}
\title{Show the Alignment to the Reference Genome}
\usage{
show_alignment(read, bam_sam_file, ref_genome)
}
\arguments{
\item{read}{Sequence of bases (ACGT)}

\item{bam_sam_file}{BAM/SAM file}

\item{ref_genome}{Reference genome from UCSC}
}
\value{
position, cigar string, read and reference sequence
}
\description{
This function reads in a BAM/SAM file, taking a specific read entry
and the reference genome and show the precise alignment to the
reference genome according to the CIGAR string of the read.
}
\examples{
library(BSgenome.Hsapiens.UCSC.hg38)
file <- system.file('extdata', 'example.bam', package="ShowAlignment")
show_alignment("GCATAACCCCTCTCTGAATCAATGAATCGCAAATCAGCCACTTTAATTAAGCTAAGCCCTTACTAGACCAATGGGACTTAAACCCGCAAACACTTAGTTAA", file, Hsapiens)

}
\references{
\url{https://en.wikipedia.org/wiki/Sequence_alignment#Representations}\cr
}
\author{
Erica Pirotta\cr Politecnico di Milano\cr Maintainer: Erica
Pirotta \cr E-Mail: \href{mailto:erica.pirotta@mail.polimi.it}{erica.pirotta@mail.polimi.it}
}
