% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showAlignment-package.R
\docType{package}
\name{showAlignment-package}
\alias{showAlignment-package}
\alias{showAlignment}
\title{showAlignment - Show the alignment of mapped reads on the reference genome}
\description{
The \code{showAlignment} package implements functions that allow to show
the precise alignment of a read entry from a BAM/SAM file to the reference genome
according to the CIGAR string of the read considered.
}
\details{
\tabular{ll}{
Package: \tab showAlignment\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2022-06-01\cr
License: \tab GPL (>=2)\cr
}
}
\references{
\url{https://en.wikipedia.org/wiki/Sequence_alignment#Representations}\cr
}
\author{
Erica Pirotta\cr
Politecnico di Milano\cr
Maintainer: Erica Pirotta\cr
E-Mail: \href{mailto:erica.pirotta@mail.polimi.it}{erica.pirotta@mail.polimi.it}
}
