% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_bam_function.R
\name{upload_bam}
\alias{upload_bam}
\title{Upload BAM file}
\usage{
upload_bam(file_bam)
}
\arguments{
\item{file_bam}{BAM file}
}
\value{
a dataframe containing the information of the BAM file
}
\description{
This function allow to scan the BAM file and to view it as a dataframe
}
\examples{
bam_file <- system.file('extdata', 'example.bam', package="ShowAlignment")
upload_bam(bam_file)

}
\references{
\url{https://en.wikipedia.org/wiki/SAM_(file_format)}\cr
}
\author{
Erica Pirotta\cr Politecnico di Milano\cr Maintainer: Erica
Pirotta \cr E-Mail: \href{mailto:erica.pirotta@mail.polimi.it}{erica.pirotta@mail.polimi.it}
}
