% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_sam_function.R
\name{upload_sam}
\alias{upload_sam}
\title{Upload SAM file}
\usage{
upload_sam(file_sam)
}
\arguments{
\item{file_sam}{SAM file}
}
\value{
a dataframe containing the information of the SAM file
}
\description{
This function allow to read a file in SAM format and to view it as a dataframe
}
\examples{
sam_file <- system.file('extdata', 'example.sam', package="ShowAlignment")
upload_sam(sam_file)

}
\references{
\url{https://en.wikipedia.org/wiki/SAM_(file_format)}\cr
}
\author{
Erica Pirotta\cr Politecnico di Milano\cr Maintainer: Erica
Pirotta \cr E-Mail: \href{mailto:erica.pirotta@mail.polimi.it}{erica.pirotta@mail.polimi.it}
}
